+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - main:  Fixed handling command line arguments from SonarQube scan finding.
 
 ### Changed
+- mongo_dump:  Replaced cmds_gen.run_prog with subprocess code to run dump command.
 - sync_cp_dump:  Add email capability for dumps error/warnings detected.
 - run_program:  Setup and configured email.
 - main:  Added -e and -s options for email capability.